// src/Interfaces/score.interface.ts
export interface I_SCORE_DETAILS {
    mps_sf_id: string;
    mps_painter_code: string;
    mps_final_painter_score: number;
    painter_status: string;
    mps_lm_adv_lead_assigned: number;
    mps_lm_lead_conversion_percentage: number;
    mps_csat_count: number;
    csat_score: number;
    conversion_score: number;
    business_score: number;
    last_updated_date: string;
    lw_points: string;
    appointment_confirmed_leads: number,
    number_of_reviews: number
}